new_requirements.txt
absl-py==0.11.0
altair==4.0.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.3
chardet==4.0.0
coloredlogs==10.0
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
docopt==0.6.2
entrypoints==0.3
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
h5py==3.1.0
humanfriendly==8.2
idna==2.10
importlib-metadata==3.1.0
importlib-resources==5.1.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.2
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab==1.0.1
jupyterlab-pygments==0.1.2
jupyterlab-server==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.5
numpy==1.16.1
packaging==20.4
pandas==0.24.2
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkg-resources==0.0.0
prometheus-client==0.9.0
promise==2.3
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.25.1
scipy==1.5.4
seaborn==0.9.0
Send2Trash==1.5.0
six==1.15.0
tensorboard==1.14.0
tensorflow-datasets==4.2.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-metadata==0.27.0
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
toolz==0.11.1
tornado==6.1
tqdm==4.24.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0